requirements.txt
streamlit>=1.18
google-generativeai>=0.4.0
qdrant-client>=1.7.0
cohere>=4.0.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
pdfplumber>=0.7.6
google-genai
PyPDF2
langchain
langchain-community
google-generativeai
cohere
python-docx
PyMuPDF
